fix: 닉네임 fallback 처리 및 공통 로직 분리 (#244)
- OAuth 닉네임이 규칙에 맞지 않을 경우 그룹 닉네임을 기본값(유저)으로 저장
- 그룹 닉네임 검증/해결 로직을 공통 유틸로 분리
